Consolidate date picker component and update styles
